+++ src/components/sections/OpenSource.tsx
@@ -8,7 +8,7 @@ import React, { useMemo } from 'react';
 import vader from 'assets/vader.webp';
 import { RiOpenSourceLine } from 'react-icons/ri';
 import { Parallax } from 'react-scroll-parallax';
-import { Default, Mobile, useMobileMode } from './Responsive';
+import { Default, Mobile, useMobileMode } from '../Responsive';
 
 export default function OpenSource() {
   const { colorScheme } = useColorScheme();
