+++ src/components/sections/contact/Contact.tsx
@@ -10,7 +10,7 @@ import { FiMail, FiPhone } from 'react-icons/fi';
 import { SiResearchgate } from 'react-icons/si';
 import details from 'assets/Details';
 import LinkCarousel from './LinkCarousel';
-import { Default, Mobile, useMobileMode } from './Responsive';
+import { Default, Mobile, useMobileMode } from '../../Responsive';
 
 export default function Contact() {
   const mobile = useMobileMode();
