+++ src/components/sections/goals/Goals.tsx
@@ -9,7 +9,7 @@ import { MdOutlineRocketLaunch } from 'react-icons/md';
 import { HiOutlineSparkles } from 'react-icons/hi2';
 import { Parallax } from 'react-scroll-parallax';
 import { animated, useSpringValue } from '@react-spring/web';
-import { Default, Mobile, useMobileMode } from './Responsive';
+import { Default, Mobile, useMobileMode } from '../../Responsive';
 import Reach from './Reach';
 
 function Grid({ sx }: {sx?: SxProps}) {
